requirements.txt
Django==2.2.5
django_environ==0.9.0
djangorestframework==3.13.1
djangorestframework_simplejwt==4.3.0
environ==1.0
requests_html==0.6.6
